refactored some tests for main Launch
