+++ tests/server/test_server_routes_populate_queues.py
@@ -9,14 +9,12 @@ from mantarray_desktop_app import CUSTOMER_ACCOUNT_ID_UUID
 from mantarray_desktop_app import MAIN_FIRMWARE_VERSION_UUID
 from mantarray_desktop_app import MANTARRAY_NICKNAME_UUID
 from mantarray_desktop_app import MANTARRAY_SERIAL_NUMBER_UUID
-from mantarray_desktop_app import PLATE_BARCODE_UUID
 from mantarray_desktop_app import REFERENCE_VOLTAGE
 from mantarray_desktop_app import REFERENCE_VOLTAGE_UUID
 from mantarray_desktop_app import server
 from mantarray_desktop_app import SLEEP_FIRMWARE_VERSION_UUID
 from mantarray_desktop_app import SOFTWARE_RELEASE_VERSION_UUID
 from mantarray_desktop_app import START_MANAGED_ACQUISITION_COMMUNICATION
-from mantarray_desktop_app import START_RECORDING_TIME_INDEX_UUID
 from mantarray_desktop_app import USER_ACCOUNT_ID_UUID
 from mantarray_desktop_app import UTC_BEGINNING_DATA_ACQUISTION_UUID
 from mantarray_desktop_app import UTC_BEGINNING_RECORDING_UUID
